feature:控制器追加权限检验

+++ app/Http/Controllers/Admin/PermissionsController.php
@@ -17,6 +17,7 @@ class PermissionsController extends Controller
 {
     public function index(Request $request, PermissionFilters $filters): JsonResponse
     {
+        $this->authorize('viewAny', Permission::class);
         $permissions = Permission::latest()
             ->filter($filters)
             ->paginate($request->input('perPage', 15));
@@ -26,17 +27,20 @@ class PermissionsController extends Controller
 
     public function store(Request $request)
     {
+        $this->authorize('create', Permission::class);
         $permission = Permission::create(['name' => $request->input('name')]);
         return $this->success(new PermissionResource($permission), '创建成功', Response::HTTP_CREATED);
     }
 
     public function edit(Permission $permission)
     {
+        $this->authorize('update', $permission);
         return $this->success(new PermissionResource($permission));
     }
 
     public function update(PermissionRequest $request, Permission $permission)
     {
+        $this->authorize('update', $permission);
         $permission->name = $request->input('name');
         $permission->update();
         return $this->success(new PermissionResource($permission));
@@ -45,12 +49,14 @@ class PermissionsController extends Controller
 
     public function destroy(Permission $permission)
     {
+        $this->authorize('delete', $permission);
         $permission->delete();
         return $this->success();
     }
 
     public function addExportTask(Request $request): JsonResponse
     {
+        $this->authorize('export', Permission::class);
         $exportTask = ExportTask::addTask('导出权限' . ExportTask::exportFilesSuffix(), 'permission', $request->all());
         ExportTaskJob::dispatch($exportTask);
         return $this->success();
