+++ app/Http/Controllers/Admin/RolesController.php
@@ -17,6 +17,7 @@ class RolesController extends Controller
 {
     public function index(Request $request, RoleFilters $filters)
     {
+        $this->authorize('viewAny', Role::class);
         $roles = Role::latest()
             ->with('permissions')
             ->filter($filters)
@@ -26,12 +27,14 @@ class RolesController extends Controller
 
     public function create(): JsonResponse
     {
+        $this->authorize('create', Role::class);
         $permissions = Permission::all(['name', 'id']);
         return $this->success(['permissions' => $permissions]);
     }
 
     public function store(RoleRequest $request)
     {
+        $this->authorize('create', Role::class);
         $permissions = $request->input('permissions');
         $role =  Role::create(['name' => $request->input('name')]);
         if (!empty($permissions)) {
@@ -44,12 +47,14 @@ class RolesController extends Controller
 
     public function edit(Role $role): JsonResponse
     {
+        $this->authorize('update', $role);
         $role->load('permissions');
         return $this->success(new RoleResource($role));
     }
 
     public function update(RoleRequest $request, Role $role): JsonResponse
     {
+        $this->authorize('update', $role);
         $role->name = $request->input('name');
         $role->syncPermissions($request->input('permissions'));
         $role->load('permissions');
@@ -59,12 +64,14 @@ class RolesController extends Controller
 
     public function destroy(Role $role): JsonResponse
     {
+        $this->authorize('delete', $role);
         $role->delete();
         return $this->success();
     }
 
     public function addExportTask(Request $request): JsonResponse
     {
+        $this->authorize('export', Role::class);
         $exportTask = ExportTask::addTask('导出角色' . ExportTask::exportFilesSuffix(), 'role', $request->all());
         ExportTaskJob::dispatch($exportTask);
         return $this->success();
